WIP - show labels and project with tasks
